% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRIDGE_CONDITION_recode.R
\name{BRIDGE_CONDITION_recode}
\alias{BRIDGE_CONDITION_recode}
\title{Recoding Bridge Condition to make it easier to understand at a glance}
\usage{
BRIDGE_CONDITION_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Bridge condition in full words
}
\description{
Recoding Bridge Condition to make it easier to understand at a glance
}
