% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANNEL_COND_061_recode.R
\name{CHANNEL_COND_061_recode}
\alias{CHANNEL_COND_061_recode}
\title{Recoding the channel and channel protection condition rating to make it understandable without the coding guide}
\usage{
CHANNEL_COND_061_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
the condition of the channel and channel protection in words
}
\description{
Recoding the channel and channel protection condition rating to make it understandable without the coding guide
}
