% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CULVERT_COND_062_recode.R
\name{CULVERT_COND_062_recode}
\alias{CULVERT_COND_062_recode}
\title{Recoding the culvert condition rating to make it understandable without the coding guide}
\usage{
CULVERT_COND_062_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
the condition of the culverts in words
}
\description{
Recoding the culvert condition rating to make it understandable without the coding guide
}
