% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECK_COND_058_recode.R
\name{DECK_COND_058_recode}
\alias{DECK_COND_058_recode}
\title{Recoding the deck condition rating to make it understandable without the coding guide}
\usage{
DECK_COND_058_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
the condition of the deck in words
}
\description{
Recoding the deck condition rating to make it understandable without the coding guide
}
