% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESIGN_LOAD_031_recode.R
\name{DESIGN_LOAD_031_recode}
\alias{DESIGN_LOAD_031_recode}
\title{Recoding design load to make it understandable without the coding guide}
\usage{
DESIGN_LOAD_031_recode(nbi_data, metric = FALSE)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}

\item{metric}{Toggle for using metric or english descriptions of the live load the structure was designed for}
}
\value{
The live load the structure was designed for
}
\description{
Recoding design load to make it understandable without the coding guide
}
