% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONAL_CLASS_026.R
\name{FUNCTIONAL_CLASS_026_recode}
\alias{FUNCTIONAL_CLASS_026_recode}
\title{Recoding Maintenance to make it understandable without the coding guide}
\usage{
FUNCTIONAL_CLASS_026_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
The Funcional Classification of the inventory road in full words
}
\description{
Recoding Maintenance to make it understandable without the coding guide
}
