% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_data.R
\name{get_country_data}
\alias{get_country_data}
\title{Extracts NBI country data from the FHWA website into the global environment for use in R.}
\usage{
get_country_data(year = 2024, non_highway = FALSE)
}
\arguments{
\item{year}{The year when the data was collected}

\item{non_highway}{Toggle for including non-highway and routes under bridges}
}
\value{
nbi csv file downloaded from the nbi website
}
\description{
Extracts NBI country data from the FHWA website into the global environment for use in R.
}
\examples{
get_country_data(2023, TRUE)
get_country_data()
}
