% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state_data.R
\name{get_state_data}
\alias{get_state_data}
\title{Extracts NBI state data from the FHWA website into the global environment.}
\usage{
get_state_data(state = "Alabama", year = 2024)
}
\arguments{
\item{state}{The state where the data was collected}

\item{year}{The year when the data was collected}
}
\value{
nbi csv file downloaded from the nbi website
}
\description{
Extracts NBI state data from the FHWA website into the global environment.
}
\examples{
get_state_data("MD", 2022)
get_state_data("Iowa", 1999)
}
