% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HISTORY_037_recode.R
\name{HISTORY_037_recode}
\alias{HISTORY_037_recode}
\title{Recoding historical significance code to make it understandable without the coding guide}
\usage{
HISTORY_037_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
The historical significance of the structure in a short sentence
}
\description{
Recoding historical significance code to make it understandable without the coding guide
}
