% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAT_016_conversion.R
\name{LAT_016_conversion}
\alias{LAT_016_conversion}
\title{Converting Latitude to decimal format}
\usage{
LAT_016_conversion(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Latitude variable in decimal degree format
}
\description{
Converting Latitude to decimal format
}
