% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAT_UND_REF_055A_recode.R
\name{LAT_UND_REF_055A_recode}
\alias{LAT_UND_REF_055A_recode}
\title{Recoding reference feature of minimum lateral underclearance to make it understandable without the coding guide}
\usage{
LAT_UND_REF_055A_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
the feature being referenced in the minimum lateral underclearance in words
}
\description{
Recoding reference feature of minimum lateral underclearance to make it understandable without the coding guide
}
