% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatLong_conversion.R
\name{LatLong_conversion}
\alias{LatLong_conversion}
\title{Converting latitude and longitude to decimal format}
\usage{
LatLong_conversion(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Latitude and longitude variables in decimal degree format
}
\description{
Converting latitude and longitude to decimal format
}
