% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LONG_017_conversion.R
\name{LONG_017_conversion}
\alias{LONG_017_conversion}
\title{Converting Longitude to decimal format}
\usage{
LONG_017_conversion(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Longitude variable in decimal degree format
}
\description{
Converting Longitude to decimal format
}
