% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAINTENANCE_021_recode.R
\name{MAINTENANCE_021_recode}
\alias{MAINTENANCE_021_recode}
\title{Recoding Maintenance to make it understandable without the coding guide}
\usage{
MAINTENANCE_021_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Maintenance responsibility in full words
}
\description{
Recoding Maintenance to make it understandable without the coding guide
}
