% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEDIAN_CODE_033_recode.R
\name{MEDIAN_CODE_033_recode}
\alias{MEDIAN_CODE_033_recode}
\title{Recoding median code to make it understandable without the coding guide}
\usage{
MEDIAN_CODE_033_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Median status of the bridge
}
\description{
Recoding median code to make it understandable without the coding guide
}
