% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAVIGATION_038_recode.R
\name{NAVIGATION_038_recode}
\alias{NAVIGATION_038_recode}
\title{Recoding navigation control status to make it understandable without the coding guide}
\usage{
NAVIGATION_038_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Status of navigation control on waterway in a short sentence
}
\description{
Recoding navigation control status to make it understandable without the coding guide
}
