% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPEN_CLOSED_POSTED_041_recode.R
\name{OPEN_CLOSED_POSTED_041_recode}
\alias{OPEN_CLOSED_POSTED_041_recode}
\title{Recoding operational status to make it understandable without the coding guide}
\usage{
OPEN_CLOSED_POSTED_041_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Operational status of a structure in a short sentence
}
\description{
Recoding operational status to make it understandable without the coding guide
}
