% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPR_RATING_METH_063_recode.R
\name{OPR_RATING_METH_063_recode}
\alias{OPR_RATING_METH_063_recode}
\title{Recoding the method used to determine operating rating to make it understandable without the coding guide}
\usage{
OPR_RATING_METH_063_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
the load rating method used to determine operating rating in words
}
\description{
Recoding the method used to determine operating rating to make it understandable without the coding guide
}
