% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OWNER_022_recode.R
\name{OWNER_022_recode}
\alias{OWNER_022_recode}
\title{Recoding Owner to make it understandable without the coding guide}
\usage{
OWNER_022_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Bridge owner in full words
}
\description{
Recoding Owner to make it understandable without the coding guide
}
