% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECORD_TYPE_005A_recode.R
\name{RECORD_TYPE_005A_recode}
\alias{RECORD_TYPE_005A_recode}
\title{Recoding Record type to make it easier to understand without the coding guide}
\usage{
RECORD_TYPE_005A_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Routes going over or under the bridge insead of code
}
\description{
Recoding Record type to make it easier to understand without the coding guide
}
