% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROUTE_PREFIX_005B_recode.R
\name{ROUTE_PREFIX_005B_recode}
\alias{ROUTE_PREFIX_005B_recode}
\title{Recoding and renaming Route Prefix to make it easier to understand without the coding guide}
\usage{
ROUTE_PREFIX_005B_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
State names insead of code
}
\description{
Recoding and renaming Route Prefix to make it easier to understand without the coding guide
}
