% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATE_CODE_001_recode.R
\name{STATE_CODE_001_recode}
\alias{STATE_CODE_001_recode}
\title{Recoding and renaming State_Code to make it easier to understand without the coding guide}
\usage{
STATE_CODE_001_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
State names insead of code
}
\description{
Recoding and renaming State_Code to make it easier to understand without the coding guide
}
