% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRUCTURE_FLARED_035_recode.R
\name{STRUCTURE_FLARED_035_recode}
\alias{STRUCTURE_FLARED_035_recode}
\title{Recoding indicator of if sturcture is flared to make it understandable without the coding guide}
\usage{
STRUCTURE_FLARED_035_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Indicator of whether structure is flared in full words
}
\description{
Recoding indicator of if sturcture is flared to make it understandable without the coding guide
}
