% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRUCTURE_ON_UND_042_recode.R
\name{STRUCTURE_ON_UND_042_recode}
\alias{STRUCTURE_ON_UND_042_recode}
\title{Recoding the type of service on and under a structure to make it understandable without the coding guide}
\usage{
STRUCTURE_ON_UND_042_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
the type of service on and under the structure in words
}
\description{
Recoding the type of service on and under a structure to make it understandable without the coding guide
}
