% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRUCTURE_TYPE_043B_recode.R
\name{STRUCTURE_TYPE_043}
\alias{STRUCTURE_TYPE_043}
\title{Recoding Structure type to make it understandable without the coding guide}
\usage{
STRUCTURE_TYPE_043(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Latitude variable in decimal degree format
}
\description{
Recoding Structure type to make it understandable without the coding guide
}
