% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPR_TYPE_044_recode.R,
%   R/STRUCTURE_TYPE_043_recode.R
\name{STRUCTURE_TYPE_043_recode}
\alias{STRUCTURE_TYPE_043_recode}
\title{Recoding kind of material of the approach span and type of design to make it understandable without the coding guide}
\usage{
STRUCTURE_TYPE_043_recode(nbi_data)

STRUCTURE_TYPE_043_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Descriptions of the kind and type of the approach span in words

Descriptions of the kind and type of structure in words
}
\description{
Recoding kind of material of the approach span and type of design to make it understandable without the coding guide

Recoding structure kind of material and type of design to make it understandable without the coding guide
}
