% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUBSTRUCTURE_COND_060_recode.R
\name{SUBSTRUCTURE_COND_060_recode}
\alias{SUBSTRUCTURE_COND_060_recode}
\title{Recoding the substructure condition rating to make it understandable without the coding guide}
\usage{
SUBSTRUCTURE_COND_060_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
the condition of the substructure in words
}
\description{
Recoding the substructure condition rating to make it understandable without the coding guide
}
