% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUPERSTRUCTURE_COND_059_recode.R
\name{SUPERSTRUCTURE_COND_059_recode}
\alias{SUPERSTRUCTURE_COND_059_recode}
\title{Recoding the superstructure condition rating to make it understandable without the coding guide}
\usage{
SUPERSTRUCTURE_COND_059_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
the condition of the superstructure in words
}
\description{
Recoding the superstructure condition rating to make it understandable without the coding guide
}
