% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRAFFIC_SAFETY_FEATURES_036_recode.R
\name{TRAFFIC_SAFETY_FEATURES_036_recode}
\alias{TRAFFIC_SAFETY_FEATURES_036_recode}
\title{Recoding traffic safety feeatures to make it understandable without the coding guide}
\usage{
TRAFFIC_SAFETY_FEATURES_036_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Status of traffic safety features in a short sentence
}
\description{
Recoding traffic safety feeatures to make it understandable without the coding guide
}
