% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERT_CLR_UND_REF_054A_recode.R
\name{VERT_CLR_UND_REF_054A_recode}
\alias{VERT_CLR_UND_REF_054A_recode}
\title{Recoding reference feature of minimum vertical underclearance to make it understandable without the coding guide}
\usage{
VERT_CLR_UND_REF_054A_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
the feature being referenced in the minimum vertical underclearance in words
}
\description{
Recoding reference feature of minimum vertical underclearance to make it understandable without the coding guide
}
