% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRUCTURE_TYPE_043_recode.R
\name{STRUCTURE_TYPE_043_recode}
\alias{STRUCTURE_TYPE_043_recode}
\title{Recoding structure kind or material and type of design to make it understandable without the coding guide}
\usage{
STRUCTURE_TYPE_043_recode(nbi_data)
}
\arguments{
\item{nbi_data}{A csv file using the NBI coding system and format}
}
\value{
Descriptions of the kind and type of structure in words
}
\description{
Recoding structure kind or material and type of design to make it understandable without the coding guide
}
